Azure AI Search/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

 
azure-functions
azure-storage-blob
azure-search-documents
langchain
langchain_openai
langchain_community
PyPDF2
requests
python-docx
openai
azure-core



 
